+++ sgx/untrusted/src/ocall/protocol_ocall.c
@@ -14,9 +14,9 @@
  * setup_socket_ocall()
  ****************************************************************************/
 int setup_socket_ocall(const char *server_host,
-                       int server_host_len,
+                       size_t server_host_len,
                        const char *server_port,
-                       int server_port_len,
+                       size_t server_port_len,
                        int *socket_fd)
 {
   *socket_fd = UNSET_FD;
